Assistant checkpoint: Update sleep time storage to use minutes instead of decimals
Assistant generated file changes:
- db/schema.ts: Change averageSleep to store minutes
- server/routes.ts: Fix sleep time handling
- client/src/pages/health-stats.tsx: Update sleep time handling in UI
- db/migrations/0007_fix_sleep_storage.ts: Add migration for sleep time storage

---

User prompt:

This problem persists also there is another notable error occuring in the process.  The user has inputted a value of 2.05 but the console is trying to save an integer with the value of 2.08.  This also creates other isssues as the value is supposed to represent hours and minutes and decimals are not the optimal way to store this information.  Lets examin the UI, Server, any associated backend logic, and review the schema table.

+++ db/schema.ts
@@ -25,7 +25,7 @@ export const healthStats = pgTable("health_stats", {
   userId: integer("user_id").references(() => users.id),
   weight: integer("weight"),
   dateOfBirth: date("date_of_birth"),
-  averageSleep: numeric("average_sleep"),
+  averageSleep: integer("average_sleep"), // Stored in minutes
   profilePhotoUrl: text("profile_photo_url"),
   allergies: jsonb("allergies").$type<string[]>(),
   lastUpdated: timestamp("last_updated").default(sql`CURRENT_TIMESTAMP`),
